requirements.txt
streamlit
langchain
requests
python-dotenv
